Add plot rerender on data table selection
