+++ vis/bokeh_plots/cross_filter.py
@@ -3,10 +3,40 @@ import itertools
 from bokeh.charts import Bar, Line
 from bokeh.embed import components
 from bokeh.models.widgets import DataTable, TableColumn
-from bokeh.models import ColumnDataSource
+from bokeh.models import ColumnDataSource, CustomJS
 import numpy as np
 import pandas as pd
 from urlparse import urlparse
+
+js_callback = CustomJS(code="""
+    var data_table_ids = ['urls', 'tlds'];
+
+    setTimeout(function() { //need timeout to wait for class change
+      var global_state = {};
+      for (i=0; i<data_table_ids.length; i++) {
+        global_state[data_table_ids[i]] = get_table_state(data_table_ids[i]);
+      }
+      $.ajax({
+        type: "POST",
+        url: '/update_cross_filter_plots' + window.location.search, //session info
+        data: JSON.stringify(global_state),
+        contentType: "application/json",
+        dataType: "json",
+        success: function(response) {
+          $("#plot_area").html(response);
+        }
+      });
+    }, 10);
+
+    var get_table_state = function(id) {
+      var current = $("#".concat(id)).find(".bk-slick-cell.l0.selected");
+      var active_cells = [];
+      for (j = 0; j < current.length; j++) {
+        active_cells.push(current[j].innerText);
+      }
+      return active_cells;
+    };
+""")
 
 def create_queryframe(pages, dates):
     dates = pd.DataFrame(dates, columns=['url', 'timestamp'])
@@ -41,7 +71,8 @@ def most_common_url_table(df):
     source = ColumnDataSource(df.groupby('hostname')
                                 .count()
                                 .sort('url',ascending=False)
-                                .reset_index())
+                                .reset_index(),
+                              callback=js_callback)
     columns = [TableColumn(field="hostname", title="Site Name"),
                TableColumn(field="url", title="Count")]
     t = DataTable(source=source, columns=columns,
@@ -52,7 +83,8 @@ def site_tld_table(df):
     source = ColumnDataSource(df.groupby('tld')
                                 .count()
                                 .sort('url',ascending=False)
-                                .reset_index())
+                                .reset_index(),
+                              callback=js_callback)
     columns = [TableColumn(field="tld", title="Ending"),
                TableColumn(field="url", title="Count")]
     t = DataTable(source=source, columns=columns,
